+++ src/characters/PlasticBag.ts
@@ -1,7 +1,6 @@
 import Obstacle from "./abstract/Obstacle";
 
 class PlasticBag extends Obstacle {
-  protected readonly _isSolitary: boolean = true;
   protected readonly _imageFilename: string = "plastic-bag.png";
   protected readonly _foodValue: number = 0;
   protected readonly _damage: number = 8;