+++ src/characters/Sardine.ts
@@ -1,7 +1,6 @@
 import PackPrey from "./abstract/PackPrey";
 
 class Sardine extends PackPrey {
-  protected readonly _isSolitary: boolean = false;
   protected readonly _imageFilename: string = "sardine.png";
   protected readonly _foodValue: number = 0.6;
   protected readonly _damage: number = 0;